+++ tiles2/src/main/java/org/apache/struts/tiles2/actions/TilesAction.java
@@ -30,7 +30,12 @@ import org.apache.struts.action.ActionForm;
 import org.apache.struts.action.ActionForward;
 import org.apache.struts.action.ActionMapping;
 import org.apache.tiles.AttributeContext;
+import org.apache.tiles.TilesContainer;
 import org.apache.tiles.access.TilesAccess;
+import org.apache.tiles.request.ApplicationContext;
+import org.apache.tiles.request.Request;
+import org.apache.tiles.request.servlet.ServletRequest;
+import org.apache.tiles.request.servlet.ServletUtil;
 
 /**
  * Base class for Tiles Actions.
@@ -51,8 +56,8 @@ public abstract class TilesAction extends Action {
      *
      * @param mapping The ActionMapping used to select this instance.
      * @param form The optional ActionForm bean for this request (if any).
-     * @param request The HTTP request we are processing.
-     * @param response The HTTP response we are creating.
+     * @param req The HTTP request we are processing.
+     * @param res The HTTP response we are creating.
      *
      * @throws Exception if the application business logic throws
      *  an exception
@@ -62,14 +67,17 @@ public abstract class TilesAction extends Action {
     public ActionForward execute(
         ActionMapping mapping,
         ActionForm form,
-        HttpServletRequest request,
-        HttpServletResponse response)
+        HttpServletRequest req,
+        HttpServletResponse res)
         throws Exception {
 
         // Try to retrieve tile context
-        AttributeContext context = TilesAccess.getContainer(
-                request.getSession().getServletContext()).getAttributeContext(
-                new Object[] { request, response });
+        ApplicationContext applicationContext = ServletUtil
+                .getApplicationContext(req.getSession().getServletContext());
+        Request request = new ServletRequest(applicationContext,
+                req, res);
+        TilesContainer container = TilesAccess.getContainer(applicationContext);
+        AttributeContext context = container.getAttributeContext(request);
         if (context == null) {
             throw new ServletException(
                 "Can't find Tile context for '"
@@ -77,7 +85,7 @@ public abstract class TilesAction extends Action {
                     + "'. TilesAction subclasses must be called from a Tile");
         }
 
-        return this.execute(context, mapping, form, request, response);
+        return this.execute(context, mapping, form, req, res);
     }
 
     /**