+++ tiles2/src/main/java/org/apache/struts/tiles2/preparer/UrlPreparer.java
@@ -29,10 +29,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.apache.tiles.AttributeContext;
-import org.apache.tiles.context.TilesRequestContext;
 import org.apache.tiles.preparer.PreparerException;
 import org.apache.tiles.preparer.ViewPreparer;
-import org.apache.tiles.servlet.context.ServletTilesRequestContext;
+import org.apache.tiles.request.Request;
+import org.apache.tiles.request.servlet.ServletRequest;
 
 /**
  * @version $Rev$ $Date$
@@ -54,12 +54,12 @@ public class UrlPreparer implements ViewPreparer {
     }
 
     /** {@inheritDoc} */
-    public void execute(TilesRequestContext tilesContext,
+    public void execute(Request tilesContext,
             AttributeContext attributeContext) throws PreparerException {
 
-        if (tilesContext instanceof ServletTilesRequestContext) {
-            ServletTilesRequestContext servletTilesContext =
-                (ServletTilesRequestContext) tilesContext;
+        if (tilesContext instanceof ServletRequest) {
+            ServletRequest servletTilesContext =
+                (ServletRequest) tilesContext;
             HttpServletRequest request = servletTilesContext.getRequest();
             HttpServletResponse response = servletTilesContext.getResponse();
             RequestDispatcher rd = request.getSession().getServletContext()